SeProject/requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-sass-processor==1.2.2
django-environ==0.10.0
django-channels==0.7.0
djangorestframework==3.14.0
channels==3.0.5
environ==1.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
WhiteNoise==6.4.0
Pillow==9.4.0